Updates existing migration to improve logic

+++ src/datastudio/migrations/m230904_000000_remove_source_groups_column.php
@@ -3,8 +3,6 @@
 namespace BarrelStrength\Sprout\datastudio\migrations;
 
 use craft\db\Migration;
-use craft\db\Query;
-use craft\helpers\Json;
 
 class m230904_000000_remove_source_groups_column extends Migration
 {
@@ -12,7 +10,9 @@ class m230904_000000_remove_source_groups_column extends Migration
 
     public function safeUp(): void
     {
-        $this->dropColumn(self::DATASETS_TABLE, 'groupId');
+        if ($this->getDb()->tableExists(self::DATASETS_TABLE) && $this->getDb()->columnExists(self::DATASETS_TABLE, 'groupId')) {
+            $this->dropColumn(self::DATASETS_TABLE, 'groupId');
+        }
     }
 
     public function safeDown(): bool
